ai/fastapi-app/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
aiokafka==0.12.0
asyncmy==0.2.10
databases==0.9.0
greenlet==3.2.1
sqlalchemy==2.0.40
cryptography==44.0.2